pyproject.toml
[project]
name = "django-bank"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Smolynets",email = "oleh_smolynets@epam.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1,<6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.2.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
